# improve performance by removing use_tensor_core dependency