@merrymercy 2024-12-17
Previously, we found an accuracy regression if we remove this. Did you test the accuracy of all models?