@bjmsong 2024-12-18
result of test_eval_accuracy_large.py
| Model                          | kv-cache-dtype | eval_name | Before | After |
| ------------------------------ | -------------- | --------- | ------ | ----- |
| Yi-1.5-9B-Chat                 | auto           | mmlu      | 0.641  | 0.645 |
|                                |                | humaneval | 0.616  | 0.615 |
|                                |                | mgsm_en   | 0.868  | 0.880 |
| Meta-Llama-3.1-8B-Instruct-FP8 | fp8_e5m2       | mmlu      | 0.708  | 0.707 |
|                                |                | humaneval | 0.661  | 0.660 |
|                                |                | mgsm_en   | 0.860  | 0.852 |