@merrymercy 2024-12-26
What is your hardware?