@bjmsong 2024-12-27
A100(40GB)