@merrymercy 2024-12-27
Can you test the command in this PR https://github.com/sgl-project/sglang/pull/1511?

If it passes, we can merge this.