@bjmsong 2024-12-27
```
python -m sglang.launch_server --model meta-llama/Meta-Llama-3.1-70B-Instruct --tp 4 --enable-p2p-check

python python/sglang/test/run_eval.py --host 127.0.0.1 --port 30000 --model meta-llama/Meta-Llama-3.1-70B-Instruct --eval-name humaneval
```
Run the test on L20*4 twice, here's the result:

| Round  | Before | After |
| ------ | ------ | ----- |
| First  | 0.751  | 0.750 |
| Second | 0.748  | 0.750 |