@merrymercy 2024-12-27
Thanks for testing this! The previous bug only happens on certain hardware. We reproduced it once on 8xA100 40GB. That is why I am asking which hardware you are using. Is it possible for you to run it again on 8xA100 40GB with the exact command?

Especially @yzh119 also acknowledged it is a bug so we need to very careful here.