@bjmsong 2024-12-29
Run the test on 8xA100 40GB, here's the result:
| Round  | Before | After |
| ------ | ------ | ----- |
| First  | 0.752  | 0.750 |
| Second | 0.750  | 0.749 |
| Third  | 0.750  | 0.751 |

```
python -m sglang.launch_server --model meta-llama/Meta-Llama-3.1-70B-Instruct --tp 8 --enable-p2p-check --mem-fraction-static 0.8

python python/sglang/test/run_eval.py --host 127.0.0.1 --port 30000 --model meta-llama/Meta-Llama-3.1-70B-Instruct --eval-name humaneval
```