@merrymercy 2025-01-13
We probably won't merge this because

1. This is a known bug confirmed by https://github.com/sgl-project/sglang/pull/1511
2. You can override the decision by using https://github.com/sgl-project/sglang/blob/0bb0f7631114b8a4b614ec8f197327ea7fce645d/python/sglang/srt/layers/attention/flashinfer_backend.py#L904

Can you try to run the exact comment in #1511 again?
```
# reproduce
python3 -m sglang.launch_server --model meta-llama/Meta-Llama-3.1-70B-Instruct --tp 4 --enable-p2p-check
python3 run_eval.py --host 127.0.0.1 --port 30000 --model meta-llama/Meta-Llama-3.1-70B-Instruct --eval-name humaneval
```