@bjmsong 2025-01-17
test on A100(80GB) * 4, no accuracy degradation now
|        | humaneval |
| ------ | --------- |
| Before | 0.750     |
| After  | 0.750     |

```
# reproduce
python3 -m sglang.launch_server --model ${Meta-Llama-3.1-70B-Instruct} --tp 4 --enable-p2p-check

python3 python/sglang/test/run_eval.py --host 127.0.0.1 --port 30000 --model ${Meta-Llama-3.1-70B-Instruct} --eval-name humaneval
```